SMTP Connection `auth_password` mask with discs
closes #150

`auth_password` is security-sensitive information

That's why we should hide this attribute from user monitor
in New/Edit form and in Show page

This PR do this trick with special font `dotsfont`.

And JS handler to show/hide letters.

Do not use "input#type=password"
because New-from will be filled automatically
with admin login+password - this is wrong.

[Bugzilla](https://bugzilla.mozilla.org/show_bug.cgi?id=956906)
[Stackoverflow](https://security.stackexchange.com/questions/49326/should-websites-be-allowed-to-disable-autocomplete-on-forms-or-fields)