Fix "Create new Account" feature spec
Error:
The feature spec "Create new Account" was working with Poltergeist,
and is broken with Chrome-headless.

Background:
Chrome-webdriver fills in the form "New Account", and when fills in the
"Send balance notifications to" it somehow do not "focus out"(blur) out
of the input. And the selecting of the "Timezone" fails because the
click on "chosen" does the "focus out " from "Send balance notifications" and
not opens "Timezone"-chosen.

Solution:
After selecting the "Send balance notifications" we do one dummy click
on `<label>Timezone</label>` to emulate "focus out".
So, the next click on Timezone-chosen works correctly.