1408, move WithPayloads module from application controller into V1 Auth controller
Previously in 29bbfcfaed102777854b726302db24b1c104c8da commit for both
- Api::Rest::Customer::V1::AuthController
- Api::Rest::Admin::AuthController
was introduced new information like remote_ip for API log recording, but
instead of defining it exactly in these two controllers were defined in the Application
controller which is wrong. So this commit will fix this and move into Auth controller.