add config to control the behavior of the CallsMonitoring job
introduces new configuration options in yeti_web.yml to control
the behavior of the CallsMonitoring job when dealing with
disabled customer authentication and both origination/termination gateways:

- calls_monitoring.teardown_on_disabled_customer_auth
- calls_monitoring.teardown_on_disabled_term_gw
- calls_monitoring.teardown_on_disabled_orig_gw

The CallsMonitoring class has been updated to use these
configurations in the relevant methods. If the configs
are not present, the current flow should not be changed.