config to control the behavior of the CallsMonitoring job
introduces new configuration options in yeti_web.yml to control
the behavior of the CallsMonitoring job when dealing with
disabled customer
authentication and termination gateways:

- calls_monitoring.teardown_on_disabled_customer_auth
- calls_monitoring.teardown_on_disabled_term_gw

The CallsMonitoring class has been updated to use
these configurations in the
relevant methods. If the configs are not present,
the current behavior (terminating
calls for disabled customer auth and termination
gateways) is maintained.

The calls_monitoring_spec.rb file has been updated
with new test cases for the
introduced configurations.